+++ NEWS.md
@@ -1,15 +1,37 @@
+# tensorflow 2.13.0
+
+- `install_tensorflow()` changes:
+     - Installs TensorFlow v2.13 by default now.
+     - The `envname` argument new default is `"r-tensorflow"`. This means that 
+       unless the `envname` argument supplied, `install_tensorflow()` will now 
+       install into the `"r-tensorflow"` environment, bootstrapping a venv of 
+       that name if necessary.
+     - gains a `new_env` argument. If `TRUE`, any existing environment 
+       specified by `envname` is deleted and created anew. Defaults to `TRUE` if
+       envname is `"r-tensorflow"`, `FALSE` otherwise.
+     - If running on Linux, now detects if NVIDIA GPUs on Linux are installed, 
+       and if so, and installs cuDNN (via pip), configures symlinks for tensorflow 
+       to find cuDNN, and emits additional instructions for how to install the necessary CUDA 
+       drivers to enable GPU usage. Set new arg `configure_cudnn = FALSE` to disable.
+     - `pip_ignore_installed` default is now `FALSE` again.
+     - On Arm Macs (M1/M2), the default tensorflow package is once again installed, 
+       rather than `tensorflow-macos` and `tensorflow-metal`.
+
+- New `pillar:type_sum()` method for Tensors, giving a
+  more informative printout of Tensors in R tracebacks and tibbles.
+
 # tensorflow 2.11.0
 
 - `install_tensorflow()` now installs TF v2.11 by default.
 
 - `as_tensor()` now coerces bare R atomic vectors to R arrays before conversion.
-  As a consequence, by default, R atomic double vectors now coerce to 
+  As a consequence, by default, R atomic double vectors now coerce to
   'float64' dtype tensors instead of 'float32'.
-  
-- `shape()` gains the ability to accept vectors of length > 1 in `...`, 
+
+- `shape()` gains the ability to accept vectors of length > 1 in `...`,
   including other `tf.TensorShape`s. Shapes are automatically flattened.
-  
-- Fixed an issue where a `ListWrapper` object of trackable keras layers 
+
+- Fixed an issue where a `ListWrapper` object of trackable keras layers
   (e.g., as part of a keras model) would not convert to an R list.
 
 # tensorflow 2.9.0
@@ -17,7 +39,7 @@
 - Generic method updates:
   - New methods:
       all(), any(), sum(), prod(), min(), max(), mean(), range(),
-      cbind(), rbind(), t(), aperm(), sort(), 
+      cbind(), rbind(), t(), aperm(), sort(),
       as.vector(), as.character(), as.raster(),
       is.infinite(), is.finite(), is.nan()
   - `^` will now invoke `tf.square()` or `tf.sqrt()` directly when appropriate